solver: Do not solve system of equations using existing values
When we have a system of equations, the `is_well_defined` function
could return `true` because of values computed by previous iterations,
leading us to pick up the wrong equation for solving.

Change the heuristic to ensure that we pick up the equation with the
lowest number of other well-defined variables. To be refined.

Fixes: #1300

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>